Add files via upload
The project will provides the details about the potholes in the roads are sensed by the cameras which is placed in the front end of the vehicle. So that it can sense over a wide angle, the sensed data is send to the software application so that it can analyses the potholes with the various size and density which are in the road the fetched data is processed with the current longitude and latitude through the GPS in the vehicle. I have been using disparity map algorithm techique and CNN algorithm(Train dataset) to detect the potholes, According to the data set the Google map is updated with the various color dots. The orange color dots refers the potholes in the road. So the map has to provide an alternative optimal path to the destination is made and traces the path in the map. POTHOLE DETECTION